layout setup for match details

+++ app/src/main/java/com/nikola/jakshic/truesight/model/match/HeroHealingPerMin.java
@@ -8,7 +8,7 @@ public class HeroHealingPerMin {
 
     @SerializedName("raw")
     @Expose
-    public long raw;
+    public double raw;
     @SerializedName("pct")
     @Expose
     public double pct;
